requirements.txt
streamlit>=1.45.0
numpy>=1.26.4
matplotlib>=3.8.3
pillow>=10.2.0
